+++ sdk/translation/Azure.AI.Translation.Document/tests/DocumentTranslationClientLiveTests.cs
@@ -75,7 +75,7 @@ namespace Azure.AI.Translation.Document.Tests
         [RecordedTest]
         [TestCase(true)]
         [TestCase(false)]
-        public async Task GetTranslationsTest(bool usetokenCredential)
+        public async Task GetTranslationsStatusesTest(bool usetokenCredential)
         {
             Uri source = await CreateSourceContainerAsync(oneTestDocuments);
             Uri target = await CreateTargetContainerAsync();
@@ -85,7 +85,7 @@ namespace Azure.AI.Translation.Document.Tests
             var input = new DocumentTranslationInput(source, target, "fr");
             await client.StartTranslationAsync(input);
 
-            List<TranslationStatusResult> translations = await client.GetTranslationsAsync().ToEnumerableAsync();
+            List<TranslationStatusResult> translations = await client.GetTranslationsStatusesAsync().ToEnumerableAsync();
 
             Assert.GreaterOrEqual(translations.Count, 1);
             TranslationStatusResult oneTranslation = translations[0];